Validation de la connexion d'un client au serveur via le protocole défini (voir méthode connexion de la classe Protocole).

+++ PC2RON_Client/src/fr/pc2ron/protocole/ETypeTrame.java
@@ -24,7 +24,7 @@ public enum ETypeTrame {
             return this.type;
 	}
         
-        public static ETypeTrame getTypeTrame(byte type) {
+        public static ETypeTrame getTypeTrame(short type) {
             short typeNonSigne = (short) (type & 0xff);
             
             ETypeTrame eTypes[] = ETypeTrame.values();
